backend/requirements-dev.txt
fastapi==0.115.0
uvicorn==0.30.0
python-dotenv==1.0.0
requests==2.31.0
anthropic==0.40.0
boto3==1.35.0
pytest==8.3.0
moto==5.0.0